Move the retryables functions back to the ProofAggregatorStruct
